Backend/requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
beautifulsoup4==4.10.0
uvicorn==0.14.0
wordict==1.0.0
fastapi==0.68.0
nltk==3.6.3
requests==2.26.0
pydantic==1.8.2
stop-words==2018.7.23
googlesearch-python==1.0.1
